gesturebind/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pynput>=1.7.6
pyautogui>=0.9.54
PyYAML>=6.0
PyQt5>=5.15.9
tensorflow>=2.13.0
scikit-learn>=1.3.0
pillow>=10.0.0
pytest>=7.4.0
python-dotenv>=1.0.0